// src/shared/interfaces/IEstabelecimento.ts
export interface IEstabelecimento {
  id: number;

  razaoSocial: string;

  numCrfPj: string;

  nomeFantasia: string;

  ramoAtividade: string;

  cnpj: string;

  endereco: string;

  numero: string;

  complemento: string;

  bairro: string;

  cidade: string;

  estado: string;

  cep: string;

  fone: string;

  email: string;

  cel: string;

  createdAt: Date;

  updatedAt: Date;
}